Save to a directory instead of cwd

+++ story/story_manager.py
@@ -122,10 +122,15 @@ class Story:
 
     def save_to_storage(self):
         self.uuid = str(uuid.uuid1())
+        
+        save_path = "./saved_stories/"
+
+        if not os.path.exists(save_path):
+            os.makedirs(save_path)
 
         story_json = self.to_json()
         file_name = "story" + str(self.uuid) + ".json"
-        f = open(file_name, "w")
+        f = open(save_path + file_name, "w")
         f.write(story_json)
         f.close()
 
@@ -138,14 +143,18 @@ class Story:
         return self.uuid
 
     def load_from_storage(self, story_id):
+        save_path = "./saved_stories/"
+
+        if not os.path.exists(save_path):
+            return "Error save not found."
 
         file_name = "story" + story_id + ".json"
         cmd = "gsutil cp gs://aidungeonstories/" + file_name + " ."
         os.system(cmd)
-        exists = os.path.isfile(file_name)
+        exists = os.path.isfile(save_path + file_name)
 
         if exists:
-            with open(file_name, "r") as fp:
+            with open(save_path + file_name, "r") as fp:
                 game = json.load(fp)
             self.init_from_dict(game)
             return str(self)
